Refactored components into separate files
