+++ Client/src/components/app.jsx
@@ -3,6 +3,8 @@ import ReactDOM from 'react-dom';
 import $ from 'jquery';
 import styled, {createGlobalStyle} from 'styled-components';
 
+import AmenitiesRow from './amenitiesRow.jsx';
+
 const GlobalStyle = createGlobalStyle`
   body {
     font-family: 'Roboto', Helvetica Neue, sans-serif;
@@ -30,24 +32,7 @@ const AmenitiesDescription = styled.h4`
   word-wrap: break-word
 `;
 
-const Img = styled.div`
-  background-image: ${props => `url('${props.photo}')`};
-  background-size: 100% 100%;
-  background-repeat: no-repeat;
-  height: 210px;
-  width: 250px;
-`;
 
-const ImgWrapper = styled.div`
-  height: 250px;
-  width: 250px;
-  text-align: center;
-`;
-
-const ImgDescription = styled.p`
-  font-size: 16px;
-  font-weight: 400;
-`;
 
 const Button = styled.button`
   text-decoration-line: var(--font-link-text-decoration-line, none);
@@ -70,24 +55,6 @@ line-height: 1.375em;
 color: #484848
 `;
 
-const Amenity = ({photo, name}) => {
-  return (
-    <ImgWrapper className="col-2">
-      <Img photo={photo} className="img-fluid image"/>
-      <ImgDescription className="description">{name}</ImgDescription>
-    </ImgWrapper>
-  );
-};
-
-const AmenitiesRow = ({amenities}) => {
-  return (
-    <div className="row pt-3 amenitiesList">
-      {amenities.map((amenity, i) => 
-        <Amenity photo={amenity.photoUrl} name={amenity.name} key={i}/>
-      )}
-    </div>
-  );
-};
 
 export default class App extends React.Component {
   constructor(props) {
@@ -114,7 +81,6 @@ export default class App extends React.Component {
     });
   }
 
-
   render() {
     return (
       <div>
@@ -135,7 +101,4 @@ export default class App extends React.Component {
   }
 }
 
-export {
-  Amenity,
-  AmenitiesRow
-};
+
