+++ Client/src/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './components/app.jsx/index.js';
+import App from './components/app.jsx';
 
 ReactDOM.render(<App />, document.getElementById('app'));
